+++ psutil/arch/netbsd/proc.c
@@ -240,47 +240,6 @@ error:
 }
 
 
-int
-_psutil_pids(struct kinfo_proc2 **proc_list, size_t *proc_count) {
-    struct kinfo_proc2 *result;
-    // Declaring name as const requires us to cast it when passing it to
-    // sysctl because the prototype doesn't include the const modifier.
-    char errbuf[_POSIX2_LINE_MAX];
-    int cnt;
-    kvm_t *kd;
-    size_t mlen;
-
-    kd = kvm_openfiles(NULL, NULL, NULL, KVM_NO_FILES, errbuf);
-    if (kd == NULL) {
-        PyErr_Format(
-            PyExc_RuntimeError, "kvm_openfiles() syscall failed: %s", errbuf
-        );
-        return -1;
-    }
-
-    result = kvm_getproc2(kd, KERN_PROC_ALL, 0, sizeof(struct kinfo_proc2), &cnt);
-    if (result == NULL) {
-        PyErr_Format(PyExc_RuntimeError, "kvm_getproc2() syscall failed");
-        kvm_close(kd);
-        return -1;
-    }
-
-    *proc_count = (size_t)cnt;
-
-    mlen = cnt * sizeof(struct kinfo_proc2);
-    if ((*proc_list = malloc(mlen)) == NULL) {
-        PyErr_NoMemory();
-        kvm_close(kd);
-        return -1;
-    }
-
-    memcpy(*proc_list, result, mlen);
-    assert(*proc_list != NULL);
-    kvm_close(kd);
-    return 0;
-}
-
-
 PyObject *
 psutil_proc_cmdline(PyObject *self, PyObject *args) {
     pid_t pid;
